ml-backend/requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
tensorflow==2.15.0
xgboost==2.0.2

# Data Processing
scipy==1.11.4
imbalanced-learn==0.11.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
